map вынесен в отдельный модуль
